evevgegeg

+++ src/main/java/if3t/entities/City.java
@@ -23,10 +23,10 @@ public class City {
 	private String country;
 	
 	@Column(name = "longitude", nullable = false)
-	private float longitude;
+	private double longitude;
 	
 	@Column(name = "latitude", nullable = false)
-	private float latitude;
+	private double latitude;
 
 	public Long getId() {
 		return id;
@@ -52,19 +52,19 @@ public class City {
 		this.country = country;
 	}
 
-	public float getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(float longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 
-	public float getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(float latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 
@@ -98,6 +98,4 @@ public class City {
 			return false;
 		return true;
 	}
-	
-	
 }
